feat: Add `questions.yaml` to policies
If a policy is simple enough, try to cover all settings in
`questions.yaml`.

If not, opt out and use a question for the whole policy settings, using
a multiline string variable. No check is done to this multiline
variable, it only gets transformed to yaml in the template.

Some of them are based in the work done already done by Jordon in his
first approach.

Co-authored-by: Jordon Leach <jordonleach@gmail.com>
Co-authored-by: Víctor Cuadrado Juan <vcuadradojuan@suse.de>